index compress v2 pass
